nilm-mscat/requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.1.0
h5py>=3.7.0
tables>=3.8.0

# Audio/Signal processing
librosa>=0.10.0

# Visualization and logging
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0

# Configuration and utilities
PyYAML>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# CRF and sequence modeling
python-crfsuite>=0.9.9

# Optional: for advanced features
# transformers>=4.20.0
# einops>=0.6.0

# Development tools (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991

# For MQTT/Kafka support (online inference)
# paho-mqtt>=1.6.0
# kafka-python>=2.0.0

# For serial communication
# pyserial>=3.5